Make sure we `terraform-init` inside of `run-command`  (#5258)
## Summary

When we ran ECS tasks from the command line by calling
`./bin/run-command ...`, it wasn't actually setting the environment
correctly. In fact, it wasn't using the environment variable for
anything.

This PR makes sure we call `terraform init` before we do the run
command, so that it executes in the right environment.

It also updates the lockfile under nofos/service.

## Context for reviewers

@coilysiren and I discussed this yesterday.

## Validation steps

For me, I ran this against both nofo environments to test this:

- `./bin/run-command nofos dev '["ls"]'`
- `./bin/run-command nofos prod '["ls"]'`

Before this change, it would only ever run tasks in `nofos-dev`.