[#5065] Add missing environment variables to the NOFO Builder (#5278)
## Summary

Makes progress on #5065 

This is a small PR that adds some of our missing env vars to the NOFO
builder, in `nofos-dev` and `nofos-prod`.

## Changes proposed

Adds several missing environment variables. [Env vars used by the NOFO
Builder are documented in the
README](https://github.com/HHS/simpler-grants-pdf-builder?tab=readme-ov-file#environment-variables).

I am creating corresponding variables for them in AWS.

## Context for reviewers

This follows the same pattern as this previous PR to add an env var:
https://github.com/HHS/simpler-grants-gov/pull/5098/files