[Unticketed] Fix HASH passing to Release Publish (#5303)
## Summary

The Release Publish step tags the Docker image that was built. It was
using git rev-parse to figure out the hash to tag with, but the caller
already knows a better hash, so utilize that if it's passed in

## Changes proposed

Allow caller to tell us the git hash for tagging purposes, if not
supplied calculate it as it was already being figured out.

## Context for reviewers

Causes the image to publish with the wrong tag if the change made that
triggered the built happens outside the service folder (such as a change
to our shared Terraform).

## Validation steps
Tested locally
When passed in
<img width="968" alt="image"
src="https://github.com/user-attachments/assets/a5f3df69-2955-4327-be93-7d42f00da8bd"
/>
When not passed in
<img width="754" alt="image"
src="https://github.com/user-attachments/assets/e446e627-62a1-4001-8eae-cc669d14dfd9"
/>