[Issue #4542] Create ECS task to get SAM.gov extracts (#4865)
## Summary

Fixes #4542

## Changes proposed

Implements a new ECS task, fetch_sam_extracts, which pulls
nightly/monthly extracts from sam.gov, depending on when it runs and
what exists before it. It can be run in an idempotent fashion.
Add a table, sam_gov_extract, to log and track the download status of
monthly and daily extract files, and migration

## Context for reviewers
The `fetch_sam_extracts` task is designed to be run on a recurring
schedule:
It first checks if the monthly extract file for the current month
(released on the first Sunday) has been downloaded. If not, it fetches
it.
It then calculates if any daily extract files have been released since
the last monthly download and fetches any that are missing.
This "catch-up" logic ensures that our data stays current. Note that
this PR only covers the downloading and tracking of these extract files.
A separate task will be created later to handle the processing and
ingestion of the data within these files.

## Validation steps

See unit tests