[Issue #5280] Ensure TraceId is always set since it can't be None and it's helpful to have it (#5330)
## Changes proposed

Default the TraceId, but also fix the bug where we weren't passing in
the one we were logging.

This is now ready for review. I validated that the SES IAM error hiding
inside a Pinpoint Success will now throw.

One thing we'll want to do in a follow up ticket is not update the list
of search IDs if the send fails. Currently if the job fails to send and
then you run the job again it doesn't find anything to send because the
row is updated and there's no "changes."