[Issue #5280] Address email IAM policy on services without it enabled (#5340)
## Summary

Only set up the email_access IAM policy stuff if we have the service
enabled (and thereby the data we need to do the string interpolation)

## Changes proposed

Add counts so that we don't create the policy pieces unless we're going
to assign them to the app role. Introducing counts means we need to
tweak the access path slightly as a result.

## Testing

Did manual `terraform apply` in both frontend and api to ensure this
will work for both cases and apply cleanly.