[Issue #5252] Create an endpoint for a user to be able to fetch their organizations (#5342)
## Summary

Fixes #5252

## Changes proposed

Create `"/<uuid:user_id>/organizations` endpoint to return a user's
organizations
Add service to get user's organizations
Add tests

## Context for reviewers

We can determine the orgs a user is in by connecting through the
organization_user table to the organization table. We then just need to
return a list of the organization records which should have the same
shape as the above response already.

## Validation steps

See unit tests