[Issue #5352] Follow up for per App image hashes (#5353)
## Summary

Fixes #5352

## Changes proposed

Update the way the Makefile grabs "current git hash" so it's per App
aware.

## Context for reviewers

This failed in the cron run vulnerability scans because unlike calls
from push or release GitHub events, the cron job doesn't have the
context of the current git hash it's running under so it was falling
back to the old style of picking a hash. We had tried to avoid changing
this, to preserve any backward compatibility and minimize delta from the
template but this is a necessary code path we need to support.