[Issue #5337] Make Agency Search perform a more LIKE type match (#5357)
## Changes proposed

Use the OpenSearch prefix * indicator to get OpenSearch to treat the
search term like a prefix not a full word. Basic support for not doing
that if the user quotes any of their term.

## Context for reviewers

This is the smallest delta to accomplish the AC and allow us to evaluate
how the behavior is understood by users. More helpful implementation
might use fuzziness or incorporate some sort of logic to switch to fuzzy
when a certain word or character count is reached, apply the
prefix/fuzzy indicator to every word in the search not just the last
one, etc.