[Unticketed] Make it so the sam.gov task can be run (#5382)
## Summary

## Changes proposed
Make it so we can actually run the sam.gov job

Remove some TODOs I fixed but forgot to remove the comments

## Context for reviewers
The sam.gov job that fetches extracts needs to be added as an import
like so - so the actual job is loaded and known by Flask when we try to
run it. Without this, trying to run it, Flask is like "I have no idea
what that job is" because nothing in the import tree ever pulled in the
file.

Looking to test and run this in dev, so need the command to be found

## Validation steps
`make cmd args="task fetch-sam-extracts"` now works locally - job fails
due to configuration we have setup in terraform, but we can actually
start the job at least which is what I wanted.