[Issue #4842] Setup a process for handling pre/post population and custom validations on an application form (#5264)
## Summary

Fixes #4842

## Changes proposed
Create a utility that given an application form, runs pre-population,
post-population, and validation rules on it based on a configuration.

## Context for reviewers
I added a README that explains some of this, let me know if there are
any other questions you have or more detail I should add. The README
also details some things not currently handled in this iteration
(whether we need to handle some of the later ideas depends on the forms
we need to support - so unknown for now).

The recursive configuration approach means adding new rules only
requires changes to the mappers we have defined + a function to handle
it. For most field population that should be very simple. Validation
will depend on the scenario.


## Validation steps
Mainly the tests at the moment, this isn't yet hooked up to any API or
process outside the tests yet.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>