[Unticketed] Make JSON Schema required checks have right path (#5416)
## Summary

## Changes proposed
Adjust the JSONSchema library usage to add the field name to the path we
produce in our error response

## Context for reviewers
This has been a known gap in the jsonschema library for 10+ years, and
unfortunately it seems like we need to work around it by building our
own custom validator function (it's just the existing required function
with one line added).

## Validation steps
Added some tests, had to fix a few tests that were checking the old
path.