[Issue #4661] Parse sam.gov entity file and load records into the sam.gov entity table (#5368)
## Summary

Fixes #4661

## Changes proposed
Parse sam.gov entity zip file (monthly or daily) and load records into
our sam.gov entity table.

Make UEI unique in the sam.gov entity table (confirmed that it's unique
in the equivalent prod table of grants.gov)

## Context for reviewers
I originally planned for this ticket to just encompass the parsing, and
leave the file management/data load to later, BUT the way I tested this
(see validation steps) meant it was easier to just do it all together
which helped better understand the data.

This job will run after a separate job fetches the extracts from sam.gov
and puts them in the DB, so effectively we have a queue of files to
process (although most days it'll be 1) in a specific order.

I made sure to thoroughly document all of the details of how fields work
in this file for future reference - as having to refer to the
spreadsheet (can be found on the ticket) describing the mapping of the
file is very annoying.

As this is a lot of data, and a lot of stuff that could go wrong, I made
sure to add a lot of logs and metrics to be safe, especially because I
think we'll run into edge cases I can't have considered when we get
access to real data.

There is one thing I didn't fully implement and that's
deleted/deactivated + expired records in the daily extract. I have a
follow-up ticket to do that + discuss in refinement soon.

## Validation steps
You can run this script locally. First, download the example extract ZIP
from
https://open.gsa.gov/api/sam-entity-extracts-api/#sample-extract-files -
you want the FOUO Monthly file.

Run `make console` and in that create an extract record:
```py
f.SamExtractFileFactory.create(s3_path="/Users/michaelchouinard/workspace/data/SAM_FOUO_MONTHLY.zip")
```
Set the path to wherever you downloaded the file - you probably need to
copy it to be in the same folder as the code itself.

You can run this code by running `make cmd args="task
process-sam-extracts"` which will fetch that file, and process it. This
file contains more than 70k records, and I used it to test a few
assumptions (which fields can be null, which can't) as well as adjust a
few queries to improve performance significantly. I can load that file
in about 13 seconds, while earlier iterations of my implementation took
several minutes.