[Issue #5286] Enable mTLS on a SOAP Proxy/Router specific ALB (#5375)
Fixes #5286

## Summary
We need an ALB that has mutual TLS Client Certificate Passthrough on for
the SOAP Proxy/Router but that prompts browsers to ask for a client
certificate.

## Changes proposed

Move the SOAP Proxy/Router off to it's own ALB that is configured almost
identically, but supports mTLS passthrough. This also gives us the
flexibility to use a unique DNS name for this and split it's traffic
load off to other containers in the future if needed.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->

## Validation steps

Ran the terraform apply locally to validate everything was working and
happy both in API (which should get an mTLS ALB and frontend which
should be totally unaffected by this change).