[Issue #4795] Handle deactivated/expired sam.gov entity records (#5433)
## Summary

Fixes #4795 

## Changes proposed
When processing the sam.gov entity extract, handle deactivated and
expired entity records
* Deactivated get marked as deactivated (will do something in later PRs
with that flag)
* Expired does nothing but verify we also see it as expired

## Context for reviewers
Just circling back to handling a set of cases I didn't in the first pass
at the implementation. I'm not 100% certain that the behavior here is
what we want, but I set it up to be thoroughly logged/metriced so when
we're testing it, we'll be able to know if we want to make any
adjustments.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>