[Issue #5392] Fix hash pulled for vuln scans on Release deploys (#5401)
## Summary

## Changes proposed

Be more consistent on how we find the right sha to use to build, cache,
and run vuln scans on Release deploys.

## Context for reviewers

Lower envs don't have this issue because there the rev being scanned is
always a change to the app being scanned is the last commit to the whole
repo so they match up.

## Validation steps

Ran from the branch to both Dev and Prod with the Release Tag:
https://github.com/HHS/simpler-grants-gov/actions/runs/15744639755