[Issue #5388] Add attachment ID support for list fields in validation (#5440)
## Summary

Fixes #5388

## Changes proposed
Modify the rule validator for validating attachment IDs to also handle
the value we're validating being a list of attachment IDs

## Context for reviewers
Attachment ID questions in our forms can either be a single ID, or a
list of IDs (ie. the question asks you to upload multiple files). This
handles supporting that behavior, as well as building the path to
include the index in the list. We know the SF424 that we're likely going
to need to support has at least one question like this.