[Issue #5456] Fix transform competition task. (#5447)
## Summary

Fixes #5456 

## Changes proposed

This change updates the transform logic for the competition transform
task to fix a bug.

## Context for reviewers

Before this change, the transformation was failing due to the
competition not being able to find the opportunity based on a cfda
listing entry that existed in staging db schema but not in the api
schema.

This fix ensures that the correct cfda listing is associated to the
correct competition and opportunity and will not be set if there is not
match.

## Validation steps

This validation step is handled through unit tests. 

For a sanity check, you can revert changes in the transform class and
see that the test cases will fail with
```
sqlalchemy.exc.IntegrityError: (psycopg.errors.ForeignKeyViolation) insert or update on table "competition"
```