[Issue #5428] Enforce the applicant type in the create application endpoint (#5467)
## Summary

Fixes #5428 

## Changes proposed

Add `_validate_applicant_type` in `create_application`.
Update factories
Add tests

## Context for reviewers

We want to add a rule to the create application endpoint to prevent
someone from starting an application if the competition doesn't allow
individuals/orgs and they're applying as that.

The open_to_applicants set on the competition record can have individual
and/or organization in it. These dictate what types of applicants can
apply.

We know whether someone is an individual or not when creating an
application based on whether or not they pass in an organization ID when
starting the application. The logic we add should be roughly:

if applying as an organization (org ID is set), error if organization
isn't one of the allowed applicant types
if applying as an individual (org ID is not set), error if individual
isn't one of the allowed types


## Validation steps

See updated unit tests