[Issue #5494] Fix transform competition ecs task missing opportunity (#5490)
## Summary

Fixes #5494

## Changes proposed

This ticket fixes the use case when running the transform competition
ecs task and there is an opportunity_id found by the associated cfda
listing entry but no opportunity exists for that opportunity_id.

## Context for reviewers

This was found when running the transform job in dev.

## Validation steps

Unit test for running this job and use case should pass.