[Issue #5391] Add docs for setting up dummy organizations (#5459)
## Summary

Fixes #5391

## Changes proposed
Add docs for setting up a dummy organization

## Context for reviewers
I setup the dummy organizations for this ticket (made a sam.gov entity
with everyone on the teams email as the EBIZ POC). Just adding docs for
the thing I already did.

Most folks don't have an account in dev or staging, so only a handful
were actually fully made, but as soon as they login we can run the
script again.