[Issue #4393] Backend Updates for migrating Opportunity tables to use UUID (#5192)
## Summary

Fixes #{4393}

## Changes proposed

This PR handles backend tasks and processes updates related to UUID
migration.
There were also some changes related to recent changes from main but for
the most part this is for backend tasks.

## Context for reviewers

This is an incremental change for updating Opportunity models to use
UUID. We are going to merge changes into to
[joshtonava/4393-modify-opportunity-tables-to-use-uuid-for-primary-key](https://github.com/HHS/simpler-grants-gov/tree/joshtonava/4393-modify-opportunity-tables-to-use-uuid-for-primary-key)

## Validation steps

All tests should pass

```bash
make test
```

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>