[Issue #5387] Add rule validation to API endpoints (#5443)
## Summary

Fixes #5387 

## Changes proposed
Connect the rule validation to our application and application form
logic

## Context for reviewers
The rule JSON schema we've setup can do pre/post population, and custom
validation. For the pilot, we 100% need custom validation. This ticket
enables that. It deliberately does not enable pre/post-population, but
does note where we'd have to adjust some things to support that, mostly
configuration related.

For pre/post/validation - we need to run them at certain times (eg.
post-population only on submit, OR only validation in GET endpoints) so
we actually have a few different configurations. To help select the
configuration, I made the calling methods pass in the ApplicationAction
they're doing which we then can use to determine which config is
relevant.