[Issue #5602] Update is_open property on Competition model. (#5604)
## Summary

Work for #5602

## Changes proposed

This pr separates the date check from the current implementation of
is_open property and is utilized by the soap GetOpportunityList
services.

## Context for reviewers

Currently, the GetOpportunityList operation queries simpler database and
filters out competitions that are not open.

The current logic checks both the date ranges as well the
is_simpler_grants_enabled property to determine if it is open. For the
SOAP response, we do not need to check the is_simpler_grants_enabled
property so we need to separate the date check from the current
implementation.

## Validation steps

Unit tests pass