[Issue #5520] Create the JSON schema for a SF424A (#5586)
Fixes #5520

Setup the JSON schema for an SF424A budget form

Slightly adjustment to the error message for maxItems

Adjustments to DB seed script to add this form to the local competition
+ adjusted how we setup the other competition that uses more generic
forms.

This is the complex budget form that has a whole bunch of tables, but
the data structure works a bit differently internally as line items
cross tables. I've tried my best to label the fields and where they
appear.

I adjusted the maxItems validator because the message dumped the entire
array into the message which if something like the activity line items
exceeded was just unreadable due to the size.

Something wasn't working on the frontend for me to view the forms (not
just these forms, but any), this form has almost nothing in it yet for
the UI schema as it will be very custom, so there wouldn't be anything
to view anyways.