[Issue #4393] Migrate primary keys using integers tied to legacy system to be UUID (#5621)
## Summary

Work for #4393

## Changes proposed

A series of changes are included that update all primary keys that were
integers relating to the legacy system to instead be using UUIDs.

## Context for reviewers

This work has been done incrementally with features being merged into
this branch.

- https://github.com/HHS/simpler-grants-gov/pull/5063
- https://github.com/HHS/simpler-grants-gov/pull/5118
- https://github.com/HHS/simpler-grants-gov/pull/5192
- https://github.com/HHS/simpler-grants-gov/pull/5567
- https://github.com/HHS/simpler-grants-gov/pull/5596
- https://github.com/HHS/simpler-grants-gov/pull/5619 (merged to main)


## Validation steps

All of the tests are passing in the frontend and backed.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>
Co-authored-by: Michael Chouinard <46358556+chouinar@users.noreply.github.com>