[Issue #5377] Add ability to "reset" email notifications to only pick up changes since the reset was run (#5512)
## Changes proposed

We need a config option to enable/disable the actual sending of emails
out across the wire. We still want to do the rest of the process, and
act like we sent the email, so that we "clear" a backlog of changes we
don't want to send emails for.

## Context for reviewers

In this case we want to reset because we haven't had the email job live
in Prod for the last couple months. But changes have been happening over
that time and we don't want to send a huge update email. For search
results because we need to update the list of opportunities or we'll
send the big update the next time, we want to run the job end-to-end
EXCEPT not have an email go out. Because we want to make it hard to send
emails until we've run the reset, the default is to run in reset mode
and skip sending the notification.

## Validation steps

We'll want to test the reset and email versions of the job in Dev to
validate it's all working as expected.