[Issue #5035] Utilize tinstruction data for GetOpportunityList response (#5649)
## Summary

Work for #5035

## Changes proposed

This PR queries the staging.tinstructions table to build urls for the
GetOpportunityList soap response. Specifically, this is required for
building the schema_url, and instructions_url for the opportunity.

## Context for reviewers

The tinstructions data is queried from the staging schema since we are
not transforming this data into the api schema.

## Validation steps