[Unticketed] Fix store opportunity versioning to be batched (#5653)
## Summary

## Changes proposed
Modify the store opportunity versioning task to be batched

## Context for reviewers
With the recreation of our DB in the UUID work, this task is failing
because it is running out of memory. It does need to fetch a lot of
data, we could probably adjust the query to fetch less, BUT batching it
like this should also work, and except for the initial backfill, we're
normally processing like 20 records an hour.

This general structure is identical to what I did for opportunity
attachment batching, it _should_ work to reduce memory. Even if it
doesn't, it should commit every 5k records, and the worst case scenario
would be running the job a few times.

## Validation steps
Loaded some opportunities in locally to verify batching works (with a
batch size of like 5) and will exit after 100 iterations.