[Issue #5600] Setup form for Project Abstract Summary (#5662)
## Summary

Work for #5600 

## Changes proposed
Create the form for the Project Abstract Summary

## Context for reviewers
This is a very simple form, 5 text questions. There are two fields that
in the automation that exists in grants.gov get copied from the SF424,
we don't have anything to support that planned yet, so I just noted them
down for later. Likely we can do something with our rule schema to
handle this, but that will require an entirely new rule setup to do.

## Validation steps
`make db-seed-local` will create an opportunity+competition with this
attached to it. If you navigate to it via the frontend, it looks like
this right now:

<img width="1137" height="865" alt="Screenshot 2025-07-11 at 1 37 45 PM"
src="https://github.com/user-attachments/assets/4e035ba4-4dd2-4e8c-a1ef-eedd4702fe28"
/>