[Issue #5419] Add more prepopulation rules (#5749)
## Summary

Fixes #5419

## Changes proposed
This adds additional prepopulation rules that we can configure.

## Context for reviewers
All of the newly added rules are pretty straightforward, UEI is a bit
odd for applications without orgs because they have to get a UEI still,
which is that odd `000000000INDV` one which is copied from how
grants.gov handles this.

I reworked tests and added dedicated tests for these to minimize the
complexity of the test setup a bit.