[Issue #5771] Only skip validation for is_included_in_submission during submit (#5773)
## Summary

Fixes #5771 

## Changes proposed
Adjust the logic for `is_included_in_submission` to only skip validation
when we're in the submit endpoint.

## Context for reviewers
This means validation will still happen on the application page, and the
application form page even if the flag isn't set.

The error we add for saying the `is_included_in_submission` flag is
required is only ever added during submission as well.

## Validation steps
I made a truth table for the various scenarios we want to account for
and then turned that into the
`test_validate_is_included_in_submission_behavior` test.