[Unticketed] Allow version_number to be null in search response (#5809)
## Summary

## Changes proposed
Allow `version_number` to be null in search response schema

## Context for reviewers
I setup an opportunity manually in dev/staging for our apply pilot and
set this to null in the DB, but when search tries to return it, it
errors because the schema validation. This just makes it work.

## Validation steps
Create an opportunity locally like this via `make console`
```py
x = f.OpportunityFactory()
x.current_opportunity_summary.opportunity_summary.version_number = None
dbs.commit()
```
Run `make populate-search-opportunities`

Go to localhost:3000/search and it shouldn't error.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>