[Issue #5805] Make the application_user table have a dedicated primary key ID (#5822)
## Summary

Fixes / Work for #5805 

## Changes proposed

Modify table structure of `application_user` table
Add 2 migrations
Add tests

## Context for reviewers

This table already exists, but we want to change it from having a
dual-primary key/foreign key to a dedicated primary key to make some
other tables be able to reference it a bit easier in the future. This is
exactly the same structure as the organization_user table and a similar
change.

## Validation steps

See new unit tests to confirm relationship.