[Issue #4994] Better handling of Feature Flag Defaults (#5269)
## Summary

Fixes #4994

## Changes proposed

1. We need to separate out the FF defaults from the values the user has
changed themselves so when we update the defaults we don't have old
values stuck in folks' cookies.
2. As part of that work we're standardizing on only using the middleware
to set the FF cookie values, and it will check to make sure the value
being "set" isn't the current default (it will just remove that flag
instead).
3. Instead of trying to have the FF page "react" to a user's changes
immediately we pop a message warning them they'll need to refresh the
page to apply the FF changes.
- This is because while we could make this features buttons and content
react to the FF changes, the rest of the site (including the Nav) won't
reflect those changes reliably until the page is refreshed.

## Context for reviewers
This is not an end-user facing page so we leaned into simplicity not
perfect UX or seamless rendering.

## Validation steps
1. Check out the branch
2. Visit http://localhost:3000/dev/feature-flags
3. Validate that now there are two columns for each feature, noting the
default and the current setting. This allows for easy scanning for
differences when trying to test features.
4. Click to enable/disable a feature flag.
- This should pop up the warning message that you need to refresh the
page to fully apply the flags you're changing
5. Now refresh the page.
- This should show the difference between the default and your value and
the button states should change accordingly.
6. If you test with authOn and repeat steps 4 and 5, you can see the Nav
Sign In link appear and disappear on each refresh, validating the flag
settings are working.
7. Return to Manage Feature Flags
8. Change some more flags and use the "Apply changes and return to the
App" button
9. Verify Feature Flag changes have taken effect

<img width="1212" height="805" alt="image"
src="https://github.com/user-attachments/assets/a0b2f69c-73b2-41ca-97d1-53793599f235"
/>

---------

Co-authored-by: Doug Schrashun <dougschrashun@navapbc.com>