[Issue #5314] Adjust sam.gov extract parsing based on testing (#5854)
## Summary

Fixes / Work for #5314 

## Changes proposed
Modified sam.gov extract parsing to account for duplicate UEIs in a
given extract file (see below for context)

Modified extract processing to use a bit less memory, not loading entire
entity table into memory, and processing file line-by-line instead of as
a big text blob.

A few other refactors for simplifying logic.

## Context for reviewers
I assumed that we would only receive a single row in a given extract
file for a UEI, but it turns out that was wrong. If a given entity has
multiple "EFT Indicators" then each of those will be a separate row in
the data we receive. This indicator is related to bank account
information according to the docs, and is for the most part irrelevant
to us. From validating the monthly extract, I verified that the only two
fields that ever differ across these duplicate records are EFT indicator
and cage code (a value we do not use). I also verified that every UEI
has at least one record where the EFT indicator is blank, which is the
one we will use, and skip the others (I added validation/logging in the
job to catch if this is ever not true).

Additionally, I changed some of the parsing logic because the way it was
written, we'd load the entire file, and the entire set of sam.gov
entities into the processes memory, and I know that is going to break
when we try to run it for real. Instead we fetch 10k entities at a time
(note that the monthly extract has just shy of 900k records) and the
file processing logic was restructured a bit to use proper file parsing
instead of loading the text of the file fully into memory all at once.

## Validation steps
For verifying the changes I made, I actually used extract files from
production running locally, and that data is sensitive, so can't quite
share how you can do it, but here's what I did.

I downloaded all extract files we fetched (one monthly file, 16 daily
files over the entirety of July), setup my DB to have all of them loaded
in, pointing to the files on disk. Running locally, this takes ~10
minutes, about 7.5 of which are the monthly extract. I expect it'll take
a bit longer non-locally, but we are fine with that.

I ran just the processing part, and skipped setting up files since I did
that manually.
`make cmd args="task sam-extracts --no-fetch-extracts --no-create-orgs"`

From running locally, I had zero errors processing rows, and none of the
bad scenarios occurred with data, so the assumptions I've made seem to
at least be fairly valid.