[#5314] Commit sam.gov processing more frequently (#5867)
## Summary

Work for #5314 

## Changes proposed
Instead of doing an entire extract file in one transaction, batch the
commits to the DB

## Context for reviewers
The process is running out of memory when processing a very large daily
extract non-locally. While we likely will need to bump up the memory,
this _should_ bring the memory usage down at least somewhat as
SQLAlchemy won't need to hold 800k+ records in memory before committing
to the DB. Actually measuring memory usage on a running process is
surprisingly complicated, and running with the full dataset and a memory
profiler outright crashed my laptop, so I could only really experiment
with moderate (~50k record) sized datasets. This seems to have brought
the memory down a bit, and made it a bit more consistent, but only
testing in dev is really going to say whether that's the case.

## Validation steps
I did a whole bunch of different things, but was using something called
memray to profile the process. The setup and validation is a lot of
steps that didn't provide a ton of clarity other than showing where the
memory was used, and is likely not worth trying to setup yourself.