[Unticketed] Add form schema to app form API model (#5889)
## Summary

## Changes proposed
Add form the the API schema for an application form

## Context for reviewers
I thought we already did this, but it turns out we don't return the form
from the application form endpoint. When talking with @doug-s-nava -
realized the frontend was having to do a roundabout way of getting the
form's schema on the app form page, this should help fix that weird
logic to get it via a competition.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>