[Issue #5859] Modify sam extract fetching to not refetch pending extracts (#5895)
## Summary
Work for #5859

## Changes proposed
When checking if we already fetched an extract, factor in pending
extracts

## Context for reviewers
There are 3 extract status values with the following purpose:
* `pending` - We just downloaded this - the process extract task fetches
everything in this status
* `completed` - We've processed this - the process extract task puts
everything into this after running
* `failed` - Nothing sets this status, if we had an issue processing a
file, we could move it to this status manually

Before this change it only considered a file in the `completed` status
as valid for seeing if we already handled a file. But that means if we
rerun the fetch process two times in a row, we end up creating duplicate
records in our DB which would be bad as we'd then try to process every
file twice. To protect against that, we want pending and completed
processing statuses to both keep us from fetching an extract again. Only
failed should be ignored and re-downloaded.